#83 Refactor codes for moving team member regist to independent class
