+++ whistleon-back/api-module/src/main/java/io/hala/whistleon/service/team/TeamMemberRequestService.java
@@ -1,0 +1,10 @@
+package io.hala.whistleon.service.team;
+
+import io.hala.whistleon.controller.dto.RequestTeamMemberResponseDto;
+
+public interface TeamMemberRequestService {
+
+  void registTeamMember(Long teamId);
+
+  RequestTeamMemberResponseDto findRequestTeamMembers(Long teamId);
+}
