+++ whistleon-back/api-module/src/main/java/io/hala/whistleon/service/team/TeamMemberRequestServiceImpl.java
@@ -1,0 +1,99 @@
+package io.hala.whistleon.service.team;
+
+import io.hala.whistleon.controller.dto.RequestTeamMemberDto;
+import io.hala.whistleon.controller.dto.RequestTeamMemberResponseDto;
+import io.hala.whistleon.domain.team.Team;
+import io.hala.whistleon.domain.team.TeamMemberRequest;
+import io.hala.whistleon.domain.team.TeamMemberRequestRepository;
+import io.hala.whistleon.domain.team.TeamRepository;
+import io.hala.whistleon.domain.user.User;
+import io.hala.whistleon.exception.CustomException;
+import io.hala.whistleon.exception.ExceptionCode;
+import io.hala.whistleon.service.PrincipalHelper;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@RequiredArgsConstructor
+@Service
+public class TeamMemberRequestServiceImpl implements TeamMemberRequestService {
+
+  private final PrincipalHelper principalHelper;
+  private final TeamRepository teamRepository;
+  private final TeamMemberRequestRepository teamMemberRequestRepository;
+
+  @Transactional
+  @Override
+  public void registTeamMember(Long teamId) {
+    User loginUser = principalHelper.getLoginUser();
+    if (loginUser.hasNotTeam()) {
+      Team team = getTeamById(teamId);
+      checkAlreadyExistRequest(loginUser, team);
+      TeamMemberRequest teamMemberRequest = createTeamMemberRequest(loginUser, team);
+      teamMemberRequestRepository.save(teamMemberRequest);
+    }
+  }
+
+  @Transactional
+  @Override
+  public RequestTeamMemberResponseDto findRequestTeamMembers(Long teamId) {
+    User loginUser = principalHelper.getLoginUser();
+    Team userTeam = loginUser.getTeam();
+    checkValidTeam(userTeam, teamId);
+
+    List<TeamMemberRequest> teamMemberRequests = new ArrayList<>();
+    if (loginUser.hasTeamAuth(userTeam)) {
+      teamMemberRequests = teamMemberRequestRepository.findAllByTeam(userTeam);
+    }
+
+    return this.createRequestTeamMemberResponse(teamMemberRequests);
+  }
+
+  // 아래로는 private method
+  private void checkAlreadyExistRequest(User user, Team team) {
+    TeamMemberRequest data = teamMemberRequestRepository.findByUserAndTeam(user, team)
+        .orElse(null);
+    if (data != null) {
+      throw new CustomException(ExceptionCode.DUPLICATE_TEAM_MEMBER_REQUEST);
+    }
+  }
+
+  private void checkValidTeam(Team team, Long teamId) {
+    if (team == null) {
+      throw new CustomException(ExceptionCode.HAS_NOT_TEAM);
+    }
+    if (!team.getTeamId().equals(teamId)) {
+      throw new CustomException(ExceptionCode.RESOURCES_NOT_EXIST);
+    }
+  }
+
+  private TeamMemberRequest createTeamMemberRequest(User user, Team team) {
+    return TeamMemberRequest.builder()
+        .user(user)
+        .team(team)
+        .requestDate(LocalDateTime.now())
+        .build();
+  }
+
+  private Team getTeamById(Long teamId) {
+    return teamRepository.findById(teamId)
+        .orElseThrow(() -> new CustomException(ExceptionCode.TEAM_NOT_EXIST));
+  }
+
+
+  private RequestTeamMemberResponseDto createRequestTeamMemberResponse(
+      List<TeamMemberRequest> teamMemberRequests) {
+
+    List<RequestTeamMemberDto> requestUsers = teamMemberRequests.stream()
+        .map(RequestTeamMemberDto::of)
+        .collect(Collectors.toList());
+
+    return RequestTeamMemberResponseDto.builder()
+        .requestUsers(requestUsers)
+        .build();
+  }
+}
